@@ -43,15 +43,15 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
+        <java.version>21</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
         <maven.compiler.compilerVersion>${java.version}</maven.compiler.compilerVersion>
-        <spring-boot.version>2.1.18.RELEASE</spring-boot.version>
+        <spring-boot.version>3.3.6</spring-boot.version>
         <!-- 覆盖 spring-boot-dependencies 里面的版本号 -->
-        <slf4j.version>1.7.32</slf4j.version>
-        <log4j2.version>2.17.0</log4j2.version>
-        <logback.version>1.2.9</logback.version>
+        <slf4j.version>2.0.16</slf4j.version>
+        <log4j2.version>2.24.3</log4j2.version>
+        <logback.version>1.5.12</logback.version>
     </properties>
 
     <dependencies>
