@@ -1,7 +1,7 @@
 package org.example.annotation;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.validation.AbstractBindingResult;
