@@ -1,12 +1,12 @@
 
 package org.example.dubbo.service;
 
-import com.alibaba.dubbo.config.annotation.Service;
+import org.apache.dubbo.config.annotation.DubboService;
 
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-@Service
+@DubboService
 public class HelloServiceImpl implements HelloService {
 
     @Override
