@@ -1,6 +1,7 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
 // source: Hello.proto
 
+// Protobuf Java Version: 3.25.3
 package org.example.grpc;
 
 public final class HelloProtos {
@@ -34,18 +35,10 @@ public final class HelloProtos {
       "essage\"\000B!\n\020org.example.grpcB\013HelloProto" +
       "sP\001b\006proto3"
     };
-    com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner =
-        new com.google.protobuf.Descriptors.FileDescriptor.    InternalDescriptorAssigner() {
-          public com.google.protobuf.ExtensionRegistry assignDescriptors(
-              com.google.protobuf.Descriptors.FileDescriptor root) {
-            descriptor = root;
-            return null;
-          }
-        };
-    com.google.protobuf.Descriptors.FileDescriptor
+    descriptor = com.google.protobuf.Descriptors.FileDescriptor
       .internalBuildGeneratedFileFrom(descriptorData,
         new com.google.protobuf.Descriptors.FileDescriptor[] {
-        }, assigner);
+        });
     internal_static_services_HelloMessage_descriptor =
       getDescriptor().getMessageTypes().get(0);
     internal_static_services_HelloMessage_fieldAccessorTable = new
