@@ -1,75 +1,60 @@
 package org.example.grpc;
 
 import static io.grpc.MethodDescriptor.generateFullMethodName;
-import static io.grpc.stub.ClientCalls.asyncBidiStreamingCall;
-import static io.grpc.stub.ClientCalls.asyncClientStreamingCall;
-import static io.grpc.stub.ClientCalls.asyncServerStreamingCall;
-import static io.grpc.stub.ClientCalls.asyncUnaryCall;
-import static io.grpc.stub.ClientCalls.blockingServerStreamingCall;
-import static io.grpc.stub.ClientCalls.blockingUnaryCall;
-import static io.grpc.stub.ClientCalls.futureUnaryCall;
-import static io.grpc.stub.ServerCalls.asyncBidiStreamingCall;
-import static io.grpc.stub.ServerCalls.asyncClientStreamingCall;
-import static io.grpc.stub.ServerCalls.asyncServerStreamingCall;
-import static io.grpc.stub.ServerCalls.asyncUnaryCall;
-import static io.grpc.stub.ServerCalls.asyncUnimplementedStreamingCall;
-import static io.grpc.stub.ServerCalls.asyncUnimplementedUnaryCall;
 
 /**
  */
-@javax.annotation.Generated(
-    value = "by gRPC proto compiler (version 1.12.0)",
-    comments = "Source: Hello.proto")
+@io.grpc.stub.annotations.GrpcGenerated
 public final class HelloServiceGrpc {
 
   private HelloServiceGrpc() {}
 
-  public static final String SERVICE_NAME = "services.HelloService";
+  public static final java.lang.String SERVICE_NAME = "services.HelloService";
 
   // Static method descriptors that strictly reflect the proto.
-  @io.grpc.ExperimentalApi("https://github.com/grpc/grpc-java/issues/1901")
-  @java.lang.Deprecated // Use {@link #getHelloMethod()} instead. 
-  public static final io.grpc.MethodDescriptor<org.example.grpc.HelloMessage,
-      org.example.grpc.HelloMessage> METHOD_HELLO = getHelloMethodHelper();
-
   private static volatile io.grpc.MethodDescriptor<org.example.grpc.HelloMessage,
       org.example.grpc.HelloMessage> getHelloMethod;
 
-  @io.grpc.ExperimentalApi("https://github.com/grpc/grpc-java/issues/1901")
+  @io.grpc.stub.annotations.RpcMethod(
+      fullMethodName = SERVICE_NAME + '/' + "hello",
+      requestType = org.example.grpc.HelloMessage.class,
+      responseType = org.example.grpc.HelloMessage.class,
+      methodType = io.grpc.MethodDescriptor.MethodType.UNARY)
   public static io.grpc.MethodDescriptor<org.example.grpc.HelloMessage,
       org.example.grpc.HelloMessage> getHelloMethod() {
-    return getHelloMethodHelper();
-  }
-
-  private static io.grpc.MethodDescriptor<org.example.grpc.HelloMessage,
-      org.example.grpc.HelloMessage> getHelloMethodHelper() {
     io.grpc.MethodDescriptor<org.example.grpc.HelloMessage, org.example.grpc.HelloMessage> getHelloMethod;
     if ((getHelloMethod = HelloServiceGrpc.getHelloMethod) == null) {
       synchronized (HelloServiceGrpc.class) {
         if ((getHelloMethod = HelloServiceGrpc.getHelloMethod) == null) {
-          HelloServiceGrpc.getHelloMethod = getHelloMethod = 
+          HelloServiceGrpc.getHelloMethod = getHelloMethod =
               io.grpc.MethodDescriptor.<org.example.grpc.HelloMessage, org.example.grpc.HelloMessage>newBuilder()
               .setType(io.grpc.MethodDescriptor.MethodType.UNARY)
-              .setFullMethodName(generateFullMethodName(
-                  "services.HelloService", "hello"))
+              .setFullMethodName(generateFullMethodName(SERVICE_NAME, "hello"))
               .setSampledToLocalTracing(true)
               .setRequestMarshaller(io.grpc.protobuf.ProtoUtils.marshaller(
                   org.example.grpc.HelloMessage.getDefaultInstance()))
               .setResponseMarshaller(io.grpc.protobuf.ProtoUtils.marshaller(
                   org.example.grpc.HelloMessage.getDefaultInstance()))
-                  .setSchemaDescriptor(new HelloServiceMethodDescriptorSupplier("hello"))
-                  .build();
-          }
+              .setSchemaDescriptor(new HelloServiceMethodDescriptorSupplier("hello"))
+              .build();
         }
-     }
-     return getHelloMethod;
+      }
+    }
+    return getHelloMethod;
   }
 
   /**
    * Creates a new async stub that supports all call types for the service
    */
   public static HelloServiceStub newStub(io.grpc.Channel channel) {
-    return new HelloServiceStub(channel);
+    io.grpc.stub.AbstractStub.StubFactory<HelloServiceStub> factory =
+      new io.grpc.stub.AbstractStub.StubFactory<HelloServiceStub>() {
+        @java.lang.Override
+        public HelloServiceStub newStub(io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
+          return new HelloServiceStub(channel, callOptions);
+        }
+      };
+    return HelloServiceStub.newStub(factory, channel);
   }
 
   /**
@@ -77,7 +62,14 @@ public final class HelloServiceGrpc {
    */
   public static HelloServiceBlockingStub newBlockingStub(
       io.grpc.Channel channel) {
-    return new HelloServiceBlockingStub(channel);
+    io.grpc.stub.AbstractStub.StubFactory<HelloServiceBlockingStub> factory =
+      new io.grpc.stub.AbstractStub.StubFactory<HelloServiceBlockingStub>() {
+        @java.lang.Override
+        public HelloServiceBlockingStub newStub(io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
+          return new HelloServiceBlockingStub(channel, callOptions);
+        }
+      };
+    return HelloServiceBlockingStub.newStub(factory, channel);
   }
 
   /**
@@ -85,48 +77,52 @@ public final class HelloServiceGrpc {
    */
   public static HelloServiceFutureStub newFutureStub(
       io.grpc.Channel channel) {
-    return new HelloServiceFutureStub(channel);
+    io.grpc.stub.AbstractStub.StubFactory<HelloServiceFutureStub> factory =
+      new io.grpc.stub.AbstractStub.StubFactory<HelloServiceFutureStub>() {
+        @java.lang.Override
+        public HelloServiceFutureStub newStub(io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
+          return new HelloServiceFutureStub(channel, callOptions);
+        }
+      };
+    return HelloServiceFutureStub.newStub(factory, channel);
   }
 
   /**
    */
-  public static abstract class HelloServiceImplBase implements io.grpc.BindableService {
+  public interface AsyncService {
 
     /**
      */
-    public void hello(org.example.grpc.HelloMessage request,
+    default void hello(org.example.grpc.HelloMessage request,
         io.grpc.stub.StreamObserver<org.example.grpc.HelloMessage> responseObserver) {
-      asyncUnimplementedUnaryCall(getHelloMethodHelper(), responseObserver);
-    }
-
-    @java.lang.Override public final io.grpc.ServerServiceDefinition bindService() {
-      return io.grpc.ServerServiceDefinition.builder(getServiceDescriptor())
-          .addMethod(
-            getHelloMethodHelper(),
-            asyncUnaryCall(
-              new MethodHandlers<
-                org.example.grpc.HelloMessage,
-                org.example.grpc.HelloMessage>(
-                  this, METHODID_HELLO)))
-          .build();
+      io.grpc.stub.ServerCalls.asyncUnimplementedUnaryCall(getHelloMethod(), responseObserver);
     }
   }
 
   /**
+   * Base class for the server implementation of the service HelloService.
    */
-  public static final class HelloServiceStub extends io.grpc.stub.AbstractStub<HelloServiceStub> {
-    private HelloServiceStub(io.grpc.Channel channel) {
-      super(channel);
-    }
+  public static abstract class HelloServiceImplBase
+      implements io.grpc.BindableService, AsyncService {
 
-    private HelloServiceStub(io.grpc.Channel channel,
-        io.grpc.CallOptions callOptions) {
+    @java.lang.Override public final io.grpc.ServerServiceDefinition bindService() {
+      return HelloServiceGrpc.bindService(this);
+    }
+  }
+
+  /**
+   * A stub to allow clients to do asynchronous rpc calls to service HelloService.
+   */
+  public static final class HelloServiceStub
+      extends io.grpc.stub.AbstractAsyncStub<HelloServiceStub> {
+    private HelloServiceStub(
+        io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
       super(channel, callOptions);
     }
 
     @java.lang.Override
-    protected HelloServiceStub build(io.grpc.Channel channel,
-        io.grpc.CallOptions callOptions) {
+    protected HelloServiceStub build(
+        io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
       return new HelloServiceStub(channel, callOptions);
     }
 
@@ -134,52 +130,48 @@ public final class HelloServiceGrpc {
      */
     public void hello(org.example.grpc.HelloMessage request,
         io.grpc.stub.StreamObserver<org.example.grpc.HelloMessage> responseObserver) {
-      asyncUnaryCall(
-          getChannel().newCall(getHelloMethodHelper(), getCallOptions()), request, responseObserver);
+      io.grpc.stub.ClientCalls.asyncUnaryCall(
+          getChannel().newCall(getHelloMethod(), getCallOptions()), request, responseObserver);
     }
   }
 
   /**
+   * A stub to allow clients to do synchronous rpc calls to service HelloService.
    */
-  public static final class HelloServiceBlockingStub extends io.grpc.stub.AbstractStub<HelloServiceBlockingStub> {
-    private HelloServiceBlockingStub(io.grpc.Channel channel) {
-      super(channel);
-    }
-
-    private HelloServiceBlockingStub(io.grpc.Channel channel,
-        io.grpc.CallOptions callOptions) {
+  public static final class HelloServiceBlockingStub
+      extends io.grpc.stub.AbstractBlockingStub<HelloServiceBlockingStub> {
+    private HelloServiceBlockingStub(
+        io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
       super(channel, callOptions);
     }
 
     @java.lang.Override
-    protected HelloServiceBlockingStub build(io.grpc.Channel channel,
-        io.grpc.CallOptions callOptions) {
+    protected HelloServiceBlockingStub build(
+        io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
       return new HelloServiceBlockingStub(channel, callOptions);
     }
 
     /**
      */
     public org.example.grpc.HelloMessage hello(org.example.grpc.HelloMessage request) {
-      return blockingUnaryCall(
-          getChannel(), getHelloMethodHelper(), getCallOptions(), request);
+      return io.grpc.stub.ClientCalls.blockingUnaryCall(
+          getChannel(), getHelloMethod(), getCallOptions(), request);
     }
   }
 
   /**
+   * A stub to allow clients to do ListenableFuture-style rpc calls to service HelloService.
    */
-  public static final class HelloServiceFutureStub extends io.grpc.stub.AbstractStub<HelloServiceFutureStub> {
-    private HelloServiceFutureStub(io.grpc.Channel channel) {
-      super(channel);
-    }
-
-    private HelloServiceFutureStub(io.grpc.Channel channel,
-        io.grpc.CallOptions callOptions) {
+  public static final class HelloServiceFutureStub
+      extends io.grpc.stub.AbstractFutureStub<HelloServiceFutureStub> {
+    private HelloServiceFutureStub(
+        io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
       super(channel, callOptions);
     }
 
     @java.lang.Override
-    protected HelloServiceFutureStub build(io.grpc.Channel channel,
-        io.grpc.CallOptions callOptions) {
+    protected HelloServiceFutureStub build(
+        io.grpc.Channel channel, io.grpc.CallOptions callOptions) {
       return new HelloServiceFutureStub(channel, callOptions);
     }
 
@@ -187,8 +179,8 @@ public final class HelloServiceGrpc {
      */
     public com.google.common.util.concurrent.ListenableFuture<org.example.grpc.HelloMessage> hello(
         org.example.grpc.HelloMessage request) {
-      return futureUnaryCall(
-          getChannel().newCall(getHelloMethodHelper(), getCallOptions()), request);
+      return io.grpc.stub.ClientCalls.futureUnaryCall(
+          getChannel().newCall(getHelloMethod(), getCallOptions()), request);
     }
   }
 
@@ -199,10 +191,10 @@ public final class HelloServiceGrpc {
       io.grpc.stub.ServerCalls.ServerStreamingMethod<Req, Resp>,
       io.grpc.stub.ServerCalls.ClientStreamingMethod<Req, Resp>,
       io.grpc.stub.ServerCalls.BidiStreamingMethod<Req, Resp> {
-    private final HelloServiceImplBase serviceImpl;
+    private final AsyncService serviceImpl;
     private final int methodId;
 
-    MethodHandlers(HelloServiceImplBase serviceImpl, int methodId) {
+    MethodHandlers(AsyncService serviceImpl, int methodId) {
       this.serviceImpl = serviceImpl;
       this.methodId = methodId;
     }
@@ -231,6 +223,18 @@ public final class HelloServiceGrpc {
     }
   }
 
+  public static final io.grpc.ServerServiceDefinition bindService(AsyncService service) {
+    return io.grpc.ServerServiceDefinition.builder(getServiceDescriptor())
+        .addMethod(
+          getHelloMethod(),
+          io.grpc.stub.ServerCalls.asyncUnaryCall(
+            new MethodHandlers<
+              org.example.grpc.HelloMessage,
+              org.example.grpc.HelloMessage>(
+                service, METHODID_HELLO)))
+        .build();
+  }
+
   private static abstract class HelloServiceBaseDescriptorSupplier
       implements io.grpc.protobuf.ProtoFileDescriptorSupplier, io.grpc.protobuf.ProtoServiceDescriptorSupplier {
     HelloServiceBaseDescriptorSupplier() {}
@@ -254,9 +258,9 @@ public final class HelloServiceGrpc {
   private static final class HelloServiceMethodDescriptorSupplier
       extends HelloServiceBaseDescriptorSupplier
       implements io.grpc.protobuf.ProtoMethodDescriptorSupplier {
-    private final String methodName;
+    private final java.lang.String methodName;
 
-    HelloServiceMethodDescriptorSupplier(String methodName) {
+    HelloServiceMethodDescriptorSupplier(java.lang.String methodName) {
       this.methodName = methodName;
     }
 
@@ -276,7 +280,7 @@ public final class HelloServiceGrpc {
         if (result == null) {
           serviceDescriptor = result = io.grpc.ServiceDescriptor.newBuilder(SERVICE_NAME)
               .setSchemaDescriptor(new HelloServiceFileDescriptorSupplier())
-              .addMethod(getHelloMethodHelper())
+              .addMethod(getHelloMethod())
               .build();
         }
       }
